+++ lib/home page.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class MyHomePage extends StatefulWidget {
+  MyHomePage({Key key, this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Color(0xFF121212),
+      appBar: AppBar(
+        backgroundColor: Color(0xFF1DB954),
+        centerTitle: true,
+        title: Text(widget.title),
+      ),
+    );
+  }
+}
