+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'home page.dart';
 
 void main() {
   runApp(MyApp());
@@ -13,27 +14,6 @@ class MyApp extends StatelessWidget {
         primarySwatch: Color(0xFF121212),
       ),*/
       home: MyHomePage(title: 'Spotify Artist Search'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Color(0xFF1DB954),
-        title: Text(widget.title),
-      ),
     );
   }
 }
